:recycle: Refactor PacketListener
