+++ src/main/java/com/replaymod/recording/packet/PacketListener.java
@@ -1,7 +1,5 @@
 package com.replaymod.recording.packet;
 
-import com.github.steveice10.netty.buffer.PooledByteBufAllocator;
-import com.github.steveice10.packetlib.tcp.io.ByteBufNetOutput;
 import com.google.gson.Gson;
 import com.replaymod.core.ReplayMod;
 import com.replaymod.core.utils.Restrictions;
@@ -13,13 +11,10 @@ import com.replaymod.recording.ReplayModRecording;
 import com.replaymod.recording.Setting;
 import com.replaymod.recording.gui.GuiSavingReplay;
 import com.replaymod.recording.handler.ConnectionEventHandler;
-import com.replaymod.replaystudio.PacketData;
 import com.replaymod.replaystudio.data.Marker;
-import com.replaymod.replaystudio.io.ReplayOutputStream;
 import com.replaymod.replaystudio.replay.ReplayFile;
 import com.replaymod.replaystudio.replay.ReplayMetaData;
 import com.replaymod.replaystudio.us.myles.ViaVersion.api.Pair;
-import io.netty.buffer.ByteBuf;
 import io.netty.buffer.Unpooled;
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.channel.ChannelInboundHandlerAdapter;
@@ -28,12 +23,8 @@ import net.minecraft.crash.CrashReport;
 import net.minecraft.entity.Entity;
 import net.minecraft.network.IPacket;
 import net.minecraft.network.PacketBuffer;
-import net.minecraft.network.PacketDirection;
-import net.minecraft.network.ProtocolType;
-import net.minecraft.network.datasync.EntityDataManager;
 import net.minecraft.network.login.server.SCustomPayloadLoginPacket;
 import net.minecraft.network.login.server.SEnableCompressionPacket;
-import net.minecraft.network.login.server.SLoginSuccessPacket;
 import net.minecraft.network.play.server.*;
 import net.minecraft.util.text.StringTextComponent;
 import org.apache.commons.io.FilenameUtils;
@@ -51,7 +42,6 @@ import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
 
 import static com.replaymod.core.versions.MCVer.getMinecraft;
-import static com.replaymod.replaystudio.util.Utils.writeInt;
 
 public class PacketListener extends ChannelInboundHandlerAdapter {
 
@@ -61,22 +51,14 @@ public class PacketListener extends ChannelInboundHandlerAdapter {
     private final ReplayMod core;
     private final Path outputPath;
     private final ReplayFile replayFile;
+    private final ReplayMetaData metaData;
 
+    private final PacketRecorder packetRecorder;
     private final ResourcePackRecorder resourcePackRecorder;
 
     private final ExecutorService saveService = Executors.newSingleThreadExecutor();
-    private final ReplayOutputStream packetOutputStream;
-
-    private ReplayMetaData metaData;
 
     private ChannelHandlerContext context = null;
-
-    private final long startTime;
-    private long lastSentPacket;
-    private long timePassedWhilePaused;
-    private volatile boolean serverWasPaused;
-    private ProtocolType connectionState = ProtocolType.LOGIN;
-    private boolean loginPhase = true;
 
     /**
      * Used to keep track of the last metadata save job submitted to the save service and
@@ -90,108 +72,14 @@ public class PacketListener extends ChannelInboundHandlerAdapter {
         this.replayFile = replayFile;
         this.metaData = metaData;
         this.resourcePackRecorder = new ResourcePackRecorder(replayFile);
-        this.packetOutputStream = replayFile.writePacketData();
-        this.startTime = metaData.getDate();
+        this.packetRecorder = new PacketRecorder(replayFile, metaData);
 
         saveMetaData();
     }
 
-    private void saveMetaData() {
-        int id = lastSaveMetaDataId.incrementAndGet();
-        saveService.submit(() -> {
-            if (lastSaveMetaDataId.get() != id) {
-                return; // Another job has been scheduled, it will do the hard work.
-            }
-            try {
-                synchronized (replayFile) {
-                    if (ReplayMod.isMinimalMode()) {
-                        metaData.setFileFormat("MCPR");
-                        metaData.setFileFormatVersion(ReplayMetaData.CURRENT_FILE_FORMAT_VERSION);
-                        metaData.setProtocolVersion(MCVer.getProtocolVersion());
-                        metaData.setGenerator("ReplayMod in Minimal Mode");
-
-                        try (OutputStream out = replayFile.write("metaData.json")) {
-                            String json = (new Gson()).toJson(metaData);
-                            out.write(json.getBytes());
-                        }
-                    } else {
-                        replayFile.writeMetaData(MCVer.getPacketTypeRegistry(true), metaData);
-                    }
-                }
-            } catch (IOException e) {
-                logger.error("Writing metadata:", e);
-            }
-        });
-    }
-
-    public void save(IPacket packet) {
-        // If we're not on the main thread (i.e. we're on the netty thread), then we need to schedule the saving
-        // to happen on the main thread so we can guarantee correct ordering of inbound and inject packets.
-        // Otherwise, injected packets may end up further down the packet stream than they were supposed to and other
-        // inbound packets which may rely on the injected packet would behave incorrectly when played back.
-        if (!mc.isOnExecutionThread()) {
-            mc.enqueue(() -> save(packet));
-            return;
-        }
-        try {
-            if (packet instanceof SSpawnPlayerPacket) {
-                UUID uuid = ((SSpawnPlayerPacket) packet).getUniqueId();
-                Set<String> uuids = new HashSet<>(Arrays.asList(metaData.getPlayers()));
-                uuids.add(uuid.toString());
-                metaData.setPlayers(uuids.toArray(new String[uuids.size()]));
-                saveMetaData();
-            }
-
-            if (packet instanceof SEnableCompressionPacket) {
-                return; // Replay data is never compressed on the packet level
-            }
-
-            long now = System.currentTimeMillis();
-            if (serverWasPaused) {
-                timePassedWhilePaused = now - startTime - lastSentPacket;
-                serverWasPaused = false;
-            }
-            int timestamp = (int) (now - startTime - timePassedWhilePaused);
-            lastSentPacket = timestamp;
-            PacketData packetData = getPacketData(timestamp, packet);
-            saveService.submit(() -> {
-                try {
-                    if (ReplayMod.isMinimalMode()) {
-                        // Minimal mode, ReplayStudio might not know our packet ids, so we cannot use it
-                        com.github.steveice10.netty.buffer.ByteBuf packetIdBuf = PooledByteBufAllocator.DEFAULT.buffer();
-                        com.github.steveice10.netty.buffer.ByteBuf packetBuf = packetData.getPacket().getBuf();
-                        try {
-                            new ByteBufNetOutput(packetIdBuf).writeVarInt(packetData.getPacket().getId());
-
-                            int packetIdLen = packetIdBuf.readableBytes();
-                            int packetBufLen = packetBuf.readableBytes();
-                            writeInt(packetOutputStream, (int) packetData.getTime());
-                            writeInt(packetOutputStream, packetIdLen + packetBufLen);
-                            packetIdBuf.readBytes(packetOutputStream, packetIdLen);
-                            packetBuf.getBytes(packetBuf.readerIndex(), packetOutputStream, packetBufLen);
-                        } finally {
-                            packetIdBuf.release();
-                            packetBuf.release();
-                        }
-                    } else {
-                        packetOutputStream.write(packetData);
-                    }
-                } catch (IOException e) {
-                    throw new RuntimeException(e);
-                }
-            });
-
-            if (packet instanceof SLoginSuccessPacket) {
-                moveToPlayState();
-            }
-        } catch (Exception e) {
-            logger.error("Writing packet:", e);
-        }
-    }
-
     @Override
     public void channelInactive(ChannelHandlerContext ctx) {
-        metaData.setDuration((int) lastSentPacket);
+        metaData.setDuration((int) packetRecorder.getLastSentPacket());
         saveMetaData();
 
         core.runLater(() -> {
@@ -212,52 +100,15 @@ public class PacketListener extends ChannelInboundHandlerAdapter {
                 logger.error("Waiting for save service termination:", e);
             }
             try {
-                packetOutputStream.close();
+                packetRecorder.close();
             } catch (IOException e) {
                 logger.error("Failed to close packet output stream:", e);
             }
 
-            List<Pair<Path, ReplayMetaData>> outputPaths;
-            synchronized (replayFile) {
-                try {
-                    if (!MarkerProcessor.producesAnyOutput(replayFile)) {
-                        // Immediately close the saving popup, the user doesn't care about it
-                        core.runLater(guiSavingReplay::close);
-
-                        // If we crash right here, on the next start we'll prompt the user for recovery
-                        // but we don't really want that, so drop a marker file to skip recovery for this replay.
-                        Files.createFile(outputPath.resolveSibling(outputPath.getFileName() + ".no_recover"));
-
-                        // We still have the replay, so we just save it (at least for a few weeks) in case they change their mind
-                        String replayName = FilenameUtils.getBaseName(outputPath.getFileName().toString());
-                        Path rawFolder = ReplayMod.instance.getRawReplayFolder();
-                        Path rawPath = rawFolder.resolve(outputPath.getFileName());
-                        for (int i = 1; Files.exists(rawPath); i++) {
-                            rawPath = rawPath.resolveSibling(replayName + "." + i + ".mcpr");
-                        }
-                        Files.createDirectories(rawPath.getParent());
-                        replayFile.saveTo(rawPath.toFile());
-                        replayFile.close();
-                        return;
-                    }
-
-                    replayFile.save();
-                    replayFile.close();
-
-                    if (core.getSettingsRegistry().get(Setting.AUTO_POST_PROCESS) && !ReplayMod.isMinimalMode()) {
-                        outputPaths = MarkerProcessor.apply(outputPath, guiSavingReplay.getProgressBar()::setProgress);
-                    } else {
-                        outputPaths = Collections.singletonList(new Pair<>(outputPath, metaData));
-                    }
-                } catch (Exception e) {
-                    logger.error("Saving replay file:", e);
-                    CrashReport crashReport = CrashReport.makeCrashReport(e, "Saving replay file");
-                    core.runLater(() -> Utils.error(logger, VanillaGuiScreen.wrap(mc.currentScreen), crashReport, guiSavingReplay::close));
-                    return;
-                }
+            List<Pair<Path, ReplayMetaData>> outputPaths = saveReplayFile(guiSavingReplay);
+            if (outputPaths != null) {
+                core.runLater(() -> guiSavingReplay.presentRenameDialog(outputPaths));
             }
-
-            core.runLater(() -> guiSavingReplay.presentRenameDialog(outputPaths));
         }).start();
     }
 
@@ -288,7 +139,6 @@ public class PacketListener extends ChannelInboundHandlerAdapter {
                     save(resourcePackRecorder.handleResourcePack((SSendResourcePackPacket) packet));
                     return;
                 }
-
 
                 if (packet instanceof SCustomPayloadLoginPacket) {
                     PacketBuffer buffer = new PacketBuffer(Unpooled.buffer());
@@ -322,38 +172,37 @@ public class PacketListener extends ChannelInboundHandlerAdapter {
             } catch (Exception e) {
                 logger.error("Handling packet for recording:", e);
             }
-
         }
 
         super.channelRead(ctx, msg);
     }
 
-    private <T> void DataManager_set(EntityDataManager dataManager, EntityDataManager.DataEntry<T> entry) {
-        dataManager.register(entry.getKey(), entry.getValue());
-    }
+    public void save(IPacket packet) {
+        // If we're not on the main thread (i.e. we're on the netty thread), then we need to schedule the saving
+        // to happen on the main thread so we can guarantee correct ordering of inbound and inject packets.
+        // Otherwise, injected packets may end up further down the packet stream than they were supposed to and other
+        // inbound packets which may rely on the injected packet would behave incorrectly when played back.
+        if (!mc.isOnExecutionThread()) {
+            mc.enqueue(() -> save(packet));
+            return;
+        }
 
-    @SuppressWarnings("unchecked")
-    private PacketData getPacketData(int timestamp, IPacket packet) throws Exception {
-
-        Integer packetId = getPacketId(packet);
-        ByteBuf byteBuf = Unpooled.buffer();
         try {
-            packet.writePacketData(new PacketBuffer(byteBuf));
-            return new PacketData(timestamp, new com.replaymod.replaystudio.protocol.Packet(
-                    MCVer.getPacketTypeRegistry(loginPhase),
-                    packetId,
-                    com.github.steveice10.netty.buffer.Unpooled.wrappedBuffer(
-                            byteBuf.array(),
-                            byteBuf.arrayOffset(),
-                            byteBuf.readableBytes()
-                    )
-            ));
-        } finally {
-            byteBuf.release();
-
-            if (packet instanceof SCustomPayloadPlayPacket) {
-                ((SCustomPayloadPlayPacket) packet).getBufferData().release();
+            if (packet instanceof SSpawnPlayerPacket) {
+                UUID uuid = ((SSpawnPlayerPacket) packet).getUniqueId();
+                Set<String> uuids = new HashSet<>(Arrays.asList(metaData.getPlayers()));
+                uuids.add(uuid.toString());
+                metaData.setPlayers(uuids.toArray(new String[uuids.size()]));
+                saveMetaData();
             }
+
+            if (packet instanceof SEnableCompressionPacket) {
+                return; // Replay data is never compressed on the packet level
+            }
+
+            packetRecorder.saveIntoReplayFile(packet);
+        } catch (Exception e) {
+            logger.error("Writing packet:", e);
         }
     }
 
@@ -389,35 +238,79 @@ public class PacketListener extends ChannelInboundHandlerAdapter {
     }
 
     public long getCurrentDuration() {
-        return lastSentPacket;
+        return packetRecorder.getLastSentPacket();
     }
 
     public void setServerWasPaused() {
-        this.serverWasPaused = true;
+        packetRecorder.setServerWasPaused(true);
     }
 
-    private Integer getPacketId(IPacket packet) throws IOException {
-        Integer packetId = connectionState.getPacketId(PacketDirection.CLIENTBOUND, packet);
-
-        if (packetId == null) {
-            /*
-               Retrying as the state is PLAY
-               TODO: Investigate packet order and refactor to be more consistent
-            */
-            packetId = ProtocolType.PLAY.getPacketId(PacketDirection.CLIENTBOUND, packet);
-
-            if (packetId == null) {
-                throw new IOException("Unknown packet type:" + packet.getClass());
+    private void saveMetaData() {
+        int id = lastSaveMetaDataId.incrementAndGet();
+        saveService.submit(() -> {
+            if (lastSaveMetaDataId.get() != id) {
+                return; // Another job has been scheduled, it will do the hard work.
             }
+            try {
+                synchronized (replayFile) {
+                    if (ReplayMod.isMinimalMode()) {
+                        metaData.setFileFormat("MCPR");
+                        metaData.setFileFormatVersion(ReplayMetaData.CURRENT_FILE_FORMAT_VERSION);
+                        metaData.setProtocolVersion(MCVer.getProtocolVersion());
+                        metaData.setGenerator("ReplayMod in Minimal Mode");
 
-            moveToPlayState();
-        }
-
-        return packetId;
+                        try (OutputStream out = replayFile.write("metaData.json")) {
+                            String json = (new Gson()).toJson(metaData);
+                            out.write(json.getBytes());
+                        }
+                    } else {
+                        replayFile.writeMetaData(MCVer.getPacketTypeRegistry(true), metaData);
+                    }
+                }
+            } catch (IOException e) {
+                logger.error("Writing metadata:", e);
+            }
+        });
     }
 
-    private void moveToPlayState() {
-        connectionState = ProtocolType.PLAY;
-        loginPhase = false;
+    private List<Pair<Path, ReplayMetaData>> saveReplayFile(GuiSavingReplay guiSavingReplay) {
+        synchronized (replayFile) {
+            try {
+                if (!MarkerProcessor.producesAnyOutput(replayFile)) {
+                    // Immediately close the saving popup, the user doesn't care about it
+                    core.runLater(guiSavingReplay::close);
+
+                    // If we crash right here, on the next start we'll prompt the user for recovery
+                    // but we don't really want that, so drop a marker file to skip recovery for this replay.
+                    Files.createFile(outputPath.resolveSibling(outputPath.getFileName() + ".no_recover"));
+
+                    // We still have the replay, so we just save it (at least for a few weeks) in case they change their mind
+                    String replayName = FilenameUtils.getBaseName(outputPath.getFileName().toString());
+                    Path rawFolder = ReplayMod.instance.getRawReplayFolder();
+                    Path rawPath = rawFolder.resolve(outputPath.getFileName());
+                    for (int i = 1; Files.exists(rawPath); i++) {
+                        rawPath = rawPath.resolveSibling(replayName + "." + i + ".mcpr");
+                    }
+                    Files.createDirectories(rawPath.getParent());
+                    replayFile.saveTo(rawPath.toFile());
+                    replayFile.close();
+                    return null;
+                }
+
+                replayFile.save();
+                replayFile.close();
+
+                if (core.getSettingsRegistry().get(Setting.AUTO_POST_PROCESS) && !ReplayMod.isMinimalMode()) {
+                    return MarkerProcessor.apply(outputPath, guiSavingReplay.getProgressBar()::setProgress);
+                } else {
+                    return Collections.singletonList(new Pair<>(outputPath, metaData));
+                }
+            } catch (Exception e) {
+                logger.error("Saving replay file:", e);
+                CrashReport crashReport = CrashReport.makeCrashReport(e, "Saving replay file");
+                core.runLater(() -> Utils.error(logger, VanillaGuiScreen.wrap(mc.currentScreen), crashReport, guiSavingReplay::close));
+                return null;
+            }
+        }
     }
 }
